+++ packages/runtime/test/src/interpreter/opcode-list.ts
@@ -1569,11 +1569,11 @@ describe("Teal Opcodes", function () {
   describe("Transaction opcodes", function () {
     const stack = new Stack<StackElem>();
     let interpreter: Interpreter;
-    this.beforeAll(() => {
+    before(() => {
       interpreter = new Interpreter();
       interpreter.runtime = new Runtime([]);
       interpreter.runtime.ctx.tx = TXN_OBJ;
-      interpreter.tealVersion = 2; // set tealversion to latest (to support all tx fields)
+      interpreter.tealVersion = MaxTEALVersion; // set tealversion to latest (to support all tx fields)
     });
 
     describe("Txn: Common Fields", function () {
@@ -1932,6 +1932,7 @@ describe("Teal Opcodes", function () {
     describe("Txn: Application Call Transaction", function () {
       before(function () {
         interpreter.runtime.ctx.tx.type = 'appl';
+        interpreter.runtime.ctx.tx.apid = 1847;
       });
 
       it("should push txn ApplicationID to stack", function () {
@@ -1994,12 +1995,89 @@ describe("Teal Opcodes", function () {
         assert.equal(1, stack.length());
         assert.deepEqual(TXN_OBJ.apaa[0], stack.pop());
       });
+
+      // introduced in TEALv3
+      it("should push value from foreign assets array and push NumAssets", function () {
+        // should push Accounts[0] to stack
+        let op = new Txn(["Assets", "0"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.equal(BigInt(TXN_OBJ.apas[0]), stack.pop());
+
+        // should push Accounts[1] to stack
+        op = new Txn(["Assets", "1"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.equal(BigInt(TXN_OBJ.apas[1]), stack.pop());
+
+        // index 10 should be out_of_bound
+        op = new Txn(["Assets", "10"], 1, interpreter);
+        expectRuntimeError(
+          () => op.execute(stack),
+          RUNTIME_ERRORS.TEAL.INDEX_OUT_OF_BOUND
+        );
+
+        op = new Txn(["NumAssets"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(BigInt(TXN_OBJ.apas.length), stack.pop());
+      });
+
+      it("should push value from foreign applications array and push NumApplications", function () {
+        // special case: Txn.Applications[0] represents current_applications_id
+        let op = new Txn(["Applications", "0"], 1, interpreter);
+        op.execute(stack);
+
+        assert.equal(1, stack.length());
+        assert.equal(BigInt(TXN_OBJ.apid), stack.pop());
+
+        // Txn.Applications[1] should push "1st" app_id from foreign Apps (Txn.ForeignApps[0])
+        op = new Txn(["Applications", "1"], 1, interpreter);
+        op.execute(stack);
+
+        assert.equal(1, stack.length());
+        assert.equal(BigInt(TXN_OBJ.apfa[0]), stack.pop());
+
+        // index 10 should be out_of_bound
+        op = new Txn(["Applications", "10"], 1, interpreter);
+        expectRuntimeError(
+          () => op.execute(stack),
+          RUNTIME_ERRORS.TEAL.INDEX_OUT_OF_BOUND
+        );
+
+        op = new Txn(["NumApplications"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(BigInt(TXN_OBJ.apfa.length), stack.pop());
+      });
+
+      it("should push local, global uint and byte slices from state schema to stack", function () {
+        let op = new Txn(["GlobalNumUint"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.equal(BigInt(TXN_OBJ.apgs.nui), stack.pop());
+
+        op = new Txn(["GlobalNumByteSlice"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.equal(BigInt(TXN_OBJ.apgs.nbs), stack.pop());
+
+        op = new Txn(["LocalNumUint"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.equal(BigInt(TXN_OBJ.apls.nui), stack.pop());
+
+        op = new Txn(["LocalNumByteSlice"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.equal(BigInt(TXN_OBJ.apls.nbs), stack.pop());
+      });
     });
 
     describe("Gtxn", function () {
       before(function () {
         const tx = interpreter.runtime.ctx.tx;
-        const tx2 = { ...tx, fee: 2222 };
+        // a) 'apas' represents 'foreignAssets', b) 'apfa' represents 'foreignApps' (id's of foreign apps)
+        // https://developer.algorand.org/docs/reference/transactions/
+        const tx2 = { ...tx, fee: 2222, apas: [3033, 4044], apfa: [5005, 6006, 7077] };
         interpreter.runtime.ctx.gtxs = [tx, tx2];
       });
 
@@ -2038,6 +2116,39 @@ describe("Teal Opcodes", function () {
 
         assert.equal(1, stack.length());
         assert.deepEqual(TXN_OBJ.apaa[0], stack.pop());
+      });
+
+      it("should push value from assets or applications array by index from tx group", function () {
+        let op = new Gtxn(["1", "Assets", "0"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.deepEqual(3033n, stack.pop()); // first asset from 2nd tx in group
+
+        op = new Gtxn(["0", "Assets", "0"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.deepEqual(BigInt(TXN_OBJ.apas[0]), stack.pop()); // first asset from 1st tx
+
+        op = new Gtxn(["1", "NumAssets"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.deepEqual(2n, stack.pop());
+
+        op = new Gtxn(["1", "NumApplications"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.deepEqual(3n, stack.pop());
+
+        // index 0 represent tx.apid (current application id)
+        op = new Gtxn(["1", "Applications", "0"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.deepEqual(BigInt(interpreter.runtime.ctx.tx.apid), stack.pop());
+
+        op = new Gtxn(["0", "Applications", "2"], 1, interpreter);
+        op.execute(stack);
+        assert.equal(1, stack.length());
+        assert.deepEqual(BigInt(TXN_OBJ.apfa[1]), stack.pop());
       });
     });
 
@@ -2170,17 +2281,18 @@ describe("Teal Opcodes", function () {
   describe("Global Opcode", function () {
     const stack = new Stack<StackElem>();
     let interpreter: Interpreter;
-    this.beforeAll(() => {
-      // setup 1st account (to be used as sender)
-      const acc1: AccountStoreI = new AccountStore(123, { addr: elonAddr, sk: new Uint8Array(0) }); // setup test account
-      setDummyAccInfo(acc1);
 
+    // setup 1st account (to be used as sender)
+    const acc1: AccountStoreI = new AccountStore(123, { addr: elonAddr, sk: new Uint8Array(0) }); // setup test account
+    setDummyAccInfo(acc1);
+
+    before(() => {
       interpreter = new Interpreter();
       interpreter.runtime = new Runtime([acc1]);
       interpreter.runtime.ctx.tx = TXN_OBJ;
       interpreter.runtime.ctx.gtxs = [TXN_OBJ];
       interpreter.runtime.ctx.tx.apid = 1828;
-      interpreter.tealVersion = 2; // set tealversion to latest (to support all global fields)
+      interpreter.tealVersion = MaxTEALVersion; // set tealversion to latest (to support all global fields)
     });
 
     it("should push MinTxnFee to stack", function () {
@@ -2257,6 +2369,14 @@ describe("Teal Opcodes", function () {
       assert.equal(1828n, top);
     });
 
+    it("should push CreatorAddress to stack", function () {
+      const op = new Global(['CreatorAddress'], 1, interpreter);
+      op.execute(stack);
+
+      // creator of app (id = 1848) is set as elonAddr in ../mock/stateful
+      assert.deepEqual(decodeAddress(elonAddr).publicKey, stack.pop());
+    });
+
     it("should throw error if global field is not present in teal version", function () {
       interpreter.tealVersion = 1;
 
@@ -2277,6 +2397,12 @@ describe("Teal Opcodes", function () {
 
       expectRuntimeError(
         () => new Global(['CurrentApplicationID'], 1, interpreter),
+        RUNTIME_ERRORS.TEAL.UNKNOWN_GLOBAL_FIELD
+      );
+
+      interpreter.tealVersion = 2;
+      expectRuntimeError(
+        () => new Global(['CreatorAddress'], 1, interpreter),
         RUNTIME_ERRORS.TEAL.UNKNOWN_GLOBAL_FIELD
       );
     });
@@ -2839,15 +2965,16 @@ describe("Teal Opcodes", function () {
 
   describe("Balance", () => {
     const stack = new Stack<StackElem>();
+    let interpreter: Interpreter;
 
     // setup 1st account
     const acc1: AccountStoreI = new AccountStore(123, { addr: elonAddr, sk: new Uint8Array(0) }); // setup test account
     setDummyAccInfo(acc1);
 
-    let interpreter: Interpreter;
     this.beforeAll(() => {
       interpreter = new Interpreter();
-      interpreter.runtime = new Runtime([acc1]);
+      const runtime = new Runtime([acc1]);
+      interpreter.runtime = runtime; // setup runtime
 
       // setting txn object
       interpreter.runtime.ctx.tx = TXN_OBJ;